Gather all coded_frame allocations and free functions to a single place
Allocating coded_frame is what most encoders do anyway, so it makes
sense to always allocate it in a single place. This bring uniformity to
encoder behaviour and prevents applications from erroneusly accessing
this field when not allocated.

Additionally this helps isolating encoders that export information with
it, and it heavily simplifies its deprecation.

Signed-off-by: Vittorio Giovara <vittorio.giovara@gmail.com>