lavc: Add data and linesize to AVSubtitleRect
Use the new fields directly instead of the ones from AVPicture. This removes
a layer of indirection which serves no pratical purpose whatsoever, and
will help in removing AVPicture structure completely later.

Every subtitle encoder/decoder seamlessly points to the new arrays,
so it is possible to deprecate AVSubtitleRect.pict.

Signed-off-by: Vittorio Giovara <vittorio.giovara@gmail.com>