lavc: Deprecate rc_override and rc_override_count
Their usage is unclear and confusing, and exposes internal ratecontrol
functionality which the user should not have to deal with. Also they
only apply to ratecontrol-enabled encoders, that is mpegvideoenc, and as
such they should not be part of the public API.

For the rare usecase, and as a replacement, add a private option similar
to the one present in avconv, so that both command line and API use the
same syntax.

Signed-off-by: Vittorio Giovara <vittorio.giovara@gmail.com>