aac: Allow pass-through transcoding of ambisonic audio
The defacto MOV standard mandates support for PCM and AAC: only the
latter decoder overrides the channel layout passed by the container,
so let this happen only when the layout order is native (or unspecified).

On the encoding side, until full ambisonic channel layout conversion
is implemented, add a temporary layout to allow pass-through transcoding
of files tagged with this specification.

Signed-off-by: Vittorio Giovara <vittorio.giovara@gmail.com>